requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
altgraph==0.17.4
async-generator==1.10
attrs==21.4.0
black==22.1.0
Brotli==1.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==3.3.2
click==8.0.4
colorama==0.4.4
cryptography==36.0.1
et-xmlfile==1.1.0
Faker==23.0.0
geographiclib==2.0
geopy==2.4.1
grappa==1.0.1
h11==0.13.0
idna==3.3
iniconfig==2.0.0
jaraco.classes==3.3.1
keyring==24.3.0
macholib==1.16.3
more-itertools==10.2.0
mypy-extensions==0.4.3
openpyxl==3.1.5
outcome==1.1.0
packaging==23.2
pathspec==0.9.0
pillow==10.2.0
platformdirs==2.5.1
pluggy==1.4.0
prettytable==3.2.0
pycparser==2.21
pyinstaller==6.11.0
pyinstaller-hooks-contrib==2024.9
pyOpenSSL==22.0.0
PySocks==1.7.1
pytesseract==0.3.10
pytest==8.0.0
python-dateutil==2.8.2
python-dotenv==1.0.1
requests==2.31.0
selenium==4.17.2
setuptools==75.2.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
tomli==2.0.1
trio==0.20.0
trio-websocket==0.9.2
typing_extensions==4.9.0
urllib3==1.26.8
urllib3-secure-extra==0.1.0
user_agent==0.1.10
wcwidth==0.2.5
webdriver-manager==4.0.1
wsproto==1.1.0
